% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwordpress.R
\name{addtitle}
\alias{addtitle}
\title{fetch the blog title and insert it into the md file}
\usage{
addtitle(fname)
}
\arguments{
\item{fname}{input file name, md or rmd file}
}
\value{
returndes
}
\description{
description
}
\examples{
x=c(1,2,3) 
}

