% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwordpress.R
\name{blogtable}
\alias{blogtable}
\title{generate table for given dataframe}
\usage{
blogtable(x)
}
\arguments{
\item{x}{data frame}
}
\value{
returndes
}
\description{
description
}
\examples{
x=c(1,2,3) 
}

