% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwordpress.R
\name{buildbook}
\alias{buildbook}
\title{build a book with the given ids}
\usage{
buildbook(ids, name = "~/wpbook")
}
\arguments{
\item{ids}{value}
}
\value{
returndes
}
\description{
each id is a chapter
}
\examples{
x=c(1,2,3) 
}

