% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwordpress.R
\name{embed_png}
\alias{embed_png}
\title{embed png in the markdown}
\usage{
embed_png(path, dpi = NULL)
}
\arguments{
\item{path}{value}

\item{dpi}{value}
}
\value{
returndes
}
\description{
description
}
\examples{
x=c(1,2,3) 
}

