% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwordpress.R
\name{escape_latex}
\alias{escape_latex}
\title{escape special LaTeX characters}
\usage{
escape_latex(x, newlines = FALSE, spaces = FALSE)
}
\arguments{
\item{x}{value}

\item{newlines}{value}

\item{spaces}{value}
}
\value{
returndes
}
\description{
escape special LaTeX characters
}
\examples{
x=c(1,2,3) 
}

