% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwordpress.R
\name{getBlog}
\alias{getBlog}
\title{get the blog title and content for given id}
\usage{
getBlog(id)
}
\arguments{
\item{id}{value}
}
\value{
list(title=,content=)
}
\description{
the id should be one of the fetched id
}
\examples{
x=c(1,2,3) 
}

