% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwordpress.R
\name{html_doc}
\alias{html_doc}
\title{html format for rmarkdown}
\usage{
html_doc(inputfile)
}
\arguments{
\item{inputfile}{value}
}
\value{
returndes
}
\description{
specified the fig.path according to the inputfile
}
\examples{
x=c(1,2,3) 
}

