% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwordpress.R
\name{link}
\alias{link}
\title{generate link for both html and latex according to is_latex}
\usage{
link(href, name)
}
\arguments{
\item{href}{value}

\item{name}{value}
}
\value{
returndes
}
\description{
description
}
\examples{
x=c(1,2,3) 
}

