% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwordpress.R
\name{listblogIds}
\alias{listblogIds}
\title{list the blog ids for the given page}
\usage{
listblogIds(page)
}
\arguments{
\item{page}{value}
}
\value{
returndes
}
\description{
description
}
\examples{
x=c(1,2,3) 
}

