% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwordpress.R
\name{orgbook}
\alias{orgbook}
\title{orgbook}
\usage{
orgbook(ids, name = "~/orgbook", title = "", copyright = "copyright.md",
  preface = "preface.md")
}
\arguments{
\item{ids}{value}

\item{name}{value}

\item{title}{value}

\item{copyright}{value}

\item{preface}{value}
}
\value{
returndes
}
\description{
convert the blog ids to a org file
}
\examples{
x=c(1,2,3) 
}

