% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwordpress.R
\name{partial_escape_dataframe}
\alias{partial_escape_dataframe}
\title{escape data.frame column wise}
\usage{
partial_escape_dataframe(x, except = none_escape_column)
}
\arguments{
\item{x}{value}

\item{except}{value}
}
\value{
returndes
}
\description{
description
}
\examples{
x=c(1,2,3) 
}

