% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwordpress.R
\name{pdf_doc}
\alias{pdf_doc}
\title{replacement of rmarkdown::pdf_document}
\usage{
pdf_doc(inputfile)
}
\arguments{
\item{inputfile}{value}
}
\value{
returndes
}
\description{
description
}
\examples{
x=c(1,2,3) 
}

