% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwordpress.R
\name{png_meta}
\alias{png_meta}
\title{return the png meta data given the path}
\usage{
png_meta(path)
}
\arguments{
\item{path}{value}
}
\value{
returndes
}
\description{
description
}
\examples{
x=c(1,2,3) 
}

