% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwordpress.R
\name{single_pdf}
\alias{single_pdf}
\title{generate the pdf for a single blog}
\usage{
single_pdf(id, title = "")
}
\arguments{
\item{id}{value}

\item{title}{value}
}
\value{
returndes
}
\description{
description
}
\examples{
x=c(1,2,3) 
}

