% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwordpress.R
\name{tex_chapter}
\alias{tex_chapter}
\title{title}
\usage{
tex_chapter(inputfile, latex_engine = c("xelatex", "pdflatex", "lualatex"),
  code_width = 65)
}
\arguments{
\item{inputfile}{value}

\item{latex_engine}{value}

\item{code_width}{value}
}
\value{
returndes
}
\description{
description
}
\examples{
x=c(1,2,3) 
}

