% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_post.R
\name{update_post}
\alias{update_post}
\title{update_post}
\usage{
update_post(
  file,
  post_title,
  post_id = 0,
  post_status = "publish",
  output = TRUE
)
}
\arguments{
\item{file}{The path and file name of an R Markdown document.}

\item{post_title}{The title that should be used for the post.}

\item{post_id}{A post ID to update with the HTML content.}

\item{post_status}{The WordPress post status to assign. Default "publish".}

\item{output}{Whether to output post ID/link or return a response object.}
}
\value{
No return by default. Response from httr::POST request if specified.
}
\description{
Renders an R Markdown document into an HTML fragment and
publishes the HTML as a post in WordPress using the WordPress REST API.
}
\examples{
update_post(file = 'current_file.Rmd', post_title = 'My post title')
}
