% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwordpress.R
\name{updateBlog}
\alias{updateBlog}
\title{update blog with new content, do not use it now}
\usage{
updateBlog(id)
}
\arguments{
\item{id}{value}

\item{title}{value}

\item{content}{value}
}
\value{
returndes
}
\description{
description
}
\examples{
x=c(1,2,3) 
}

