% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwordpress.R
\name{blogtable}
\alias{blogtable}
\title{generate table for given dataframe}
\usage{
blogtable(x, escape = TRUE, none_escape_column = NULL)
}
\arguments{
\item{x}{value}

\item{escape}{value}

\item{none_escape_column}{which column do not need to be escaped}
}
\value{
returndes
}
\description{
description
}
\examples{
x=c(1,2,3) 
}

