% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwordpress.R
\name{buildbook}
\alias{buildbook}
\title{build a book with the given ids}
\usage{
buildbook(ids, name = "~/wpbook", title = "", copyright = "copyright.md",
  preface = "preface.md")
}
\arguments{
\item{ids}{value}

\item{name}{value}

\item{title}{value}

\item{copyright}{the copyright md file for the book
the file should be at the markdownRoot folder}

\item{preface}{: the preface md file for the book}
}
\value{
returndes
}
\description{
each id is a chapter
}
\examples{
x=c(1,2,3) 
}

