% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwordpress.R
\name{matchfile}
\alias{matchfile}
\title{find the md or rmd file for given p123}
\usage{
matchfile(x, pattern = ".(md|rmd|tbl|Rds|org)$")
}
\arguments{
\item{x}{value}
}
\value{
returndes
}
\description{
description
}
\examples{
x=c(1,2,3) 
}

